requirements.txt
streamlit
tensorflow
pillow
numpy
googletrans==4.0.0rc1